Ui自动化初步调通,支持并发执行，兼容执行，js脚本，断言，前置，后置，错误收集,变量设置，环境设置等等

+++ galloping-ui/src/main/java/com/snail/abell/dto/StepInforDto.java
@@ -28,7 +28,7 @@ public class StepInforDto {
 
     private Long elementId;
 
-    private Long counts;
+    private Integer counts;
 
-    private Long waite;
+    private Integer waite;
 }
